requirements/base.txt
boto3==1.7.70
Pillow==5.2.0
python-magic==0.4.15